Add example of embedding Python interpreter for cython
