+++ cython/build.sh
@@ -1,4 +1,5 @@
 #!/usr/bin/bash
 
-# Should use python3
-python setup.py build_ext -if 
+python setup.py build_ext -if
+# cython --embed PyDawg.cpp -o py_dawg_main.c
+# clang -std=c11 -Wall -I/usr/include/python3.6m py_dawg_main.c -o py_dawg_main -L/usr/lib -lpython3
