+++ cython/setup.py
@@ -10,7 +10,7 @@ ext = Extension("PyDawg",
                 # libraries=["dawg2"],
                 # library_dirs=["/usr/local/lib"],
                 # runtime_library_dirs=[],
-                # extra_compile_args=["-fPIC"],
+                extra_compile_args=["-fPIC"],
                 # extra_link_args=["-static"],
                 # extra_objects=["ma.o"],
                 #["boost_program_options"],
